pyproject.toml
[tool.poetry]
name = "ralph"
version = "0.1.0"
description = "Ralph - Trading Infrastructure with FastAPI, TimescaleDB, and Redis"
authors = ["GoldIB Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
structlog = "^24.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.poetry.scripts]
ralph = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
